moved  js code  for scroll effects into separate file
